requirements/base.txt
Django==1.5.5
Pillow==2.1.0
South==0.8.2
boto==2.15.0
celery==3.0.24
dj-database-url==0.2.2
django-celery==3.0.23
django-extensions==1.2.5
django-grappelli==2.4.7
django-require==1.0.3
django-storages==1.1.8
django-tastypie==0.10.0
mimeparse==0.1.3
psycopg2==2.5.1
pytz==2013.7
